+++ modules/shared/utils/type.ts
@@ -1,0 +1,4 @@
+export interface IFormatDateRange {
+  startFrom: string;
+  endTo: string;
+}
